Modifications to enable building on my system as well as a reopen last project feature
This update includes changes to the buildfiles and libraries to enable building the project in my hands with NetBeans. Making a new project and copying the build-imp-back.xml file into the nbproject folder was the quickest way to get this to compile. I toyed with making a Maven buildfile, but was afraid that making too many changes would make it difficult to track changes from the original branch from Norbert.

Often when debugging, it is useful to be able to get back to where you were in a project quickly. Also, sometimes one just needs to restart ImageJ or your computer.

Regardless, one of the handiest modifications I ever made was the inclusion of this Reopen Last Project command.